Interpolate char controller

+++ kool-demo/src/commonMain/kotlin/de/fabmax/kool/demo/physics/terrain/PhysicsObjects.kt
@@ -1,7 +1,10 @@
 package de.fabmax.kool.demo.physics.terrain
 
 import de.fabmax.kool.KoolContext
-import de.fabmax.kool.math.*
+import de.fabmax.kool.math.PoseF
+import de.fabmax.kool.math.QuatF
+import de.fabmax.kool.math.Vec3f
+import de.fabmax.kool.math.deg
 import de.fabmax.kool.physics.*
 import de.fabmax.kool.physics.geometry.BoxGeometry
 import de.fabmax.kool.physics.geometry.PlaneGeometry
@@ -46,7 +49,7 @@ class PhysicsObjects(mainScene: Scene, terrain: Terrain, trees: Trees, ctx: Kool
         // spawn player
         playerController = PlayerController(this, mainScene, ctx).apply {
             // set spawn position
-            controller.position = Vec3d(-146.5, 47.8, -89.0)
+            controller.position = Vec3f(-146.5f, 47.8f, -89.0f)
         }
 
         world.physicsStepListeners += playerController
